requirements.txt
openai-async

openai==0.28

discord.py